Add B2B invoice payment operations and fix naming conventions
- Add invoice payment operations: Get, Create (offline), UpdateStatus
- Split B2BInvoicePayment types into separate files per type
- Rename openBalance → OpenBalance, originalBalance → OriginalBalance (Pascal Case)
- Use pattern matching in tests per code review feedback
