+++ Fusionary.BigCommerce/B2B/Types/B2BInvoicePaymentCreateResponse.cs
@@ -1,0 +1,10 @@
+namespace Fusionary.BigCommerce.B2B.Types;
+
+/// <summary>
+/// Response from creating an offline payment.
+/// </summary>
+public record B2BInvoicePaymentCreateResponse
+{
+    [JsonPropertyName("paymentId")]
+    public int PaymentId { get; init; }
+}
