+++ Fusionary.BigCommerce/B2B/Types/B2BInvoicePaymentStatusUpdate.cs
@@ -1,0 +1,10 @@
+namespace Fusionary.BigCommerce.B2B.Types;
+
+/// <summary>
+/// Request to update payment processing status.
+/// </summary>
+public record B2BInvoicePaymentStatusUpdate
+{
+    [JsonPropertyName("processingStatus")]
+    public int ProcessingStatus { get; init; }
+}
